[Feat][utils]: show suggestion in the middle of the text
Adds to Issue: #20

There is not change in how partial_match used to work for previously
for cursor being at the end of line.
If the cursor is in the middle of the text then we consider a prefix and
suffix, for which we just check if the suffix is part of suggestion or
not.

Since all are done we just need to develop capture suggestions to handle
this text in the middle usecase